Add slice native function
